Improve end game detection
Add direction to the teronimo. When a new tetronimo is selected it should
always face "up". In addition, the cells within the matrix should start
at the top row. Thus ensures that when the new shape is first drawn on
the screen is always starts at the top row of the grid.

The end game detection routine then works a lot better -- i.e. it works
as you'd expect. Previously it would detect the end game too early if
the tetronimo started rotated away from up.